test(gateway): share startup fixtures for canary maintenance proof
