+++ src/gateway/server-startup-early.test.ts
@@ -1,7 +1,18 @@
 /**
  * Early gateway startup helper tests.
  */
-import { beforeEach, describe, expect, it, vi } from "vitest";
+import { afterEach, beforeEach, describe, expect, it, vi } from "vitest";
+import { drainGlobalSingletonLifecycleState } from "../shared/global-singleton.js";
+import { getDetachedTaskLifecycleRuntime } from "../tasks/detached-task-runtime.js";
+import { getTaskById } from "../tasks/task-registry.js";
+import { createTaskFixture } from "../tasks/task-registry.test-support.js";
+import {
+  resetDetachedTaskLifecycleRuntimeForTests,
+  resetTaskFlowRegistryForTests,
+  resetTaskRegistryForTests,
+  setDetachedTaskLifecycleRuntime,
+} from "../tasks/task-runtime.test-helpers.js";
+import { withStateDirEnv } from "../test-helpers/state-dir-env.js";
 import { createGatewayPluginRuntimeGeneration } from "./server-plugin-runtime-generation.js";
 import { runGatewayCloseSteps } from "./server-shutdown.js";
 import { createGatewayMaintenanceStateForTest } from "./test-helpers.maintenance-state.js";
@@ -50,11 +61,13 @@ vi.mock("../agents/context.js", () => ({
   ensureContextWindowCacheLoaded: mocks.ensureContextWindowCacheLoaded,
 }));
 
-vi.mock("../tasks/runtime-internal.js", () => ({
+vi.mock("../tasks/runtime-internal.js", async (importOriginal) => ({
+  ...(await importOriginal<typeof import("../tasks/runtime-internal.js")>()),
   ensureTaskRuntimeStateReady: mocks.ensureTaskRuntimeStateReady,
 }));
 
-vi.mock("../tasks/task-registry.maintenance.js", () => ({
+vi.mock("../tasks/task-registry.maintenance.js", async (importOriginal) => ({
+  ...(await importOriginal<typeof import("../tasks/task-registry.maintenance.js")>()),
   configureTaskRegistryMaintenance: mocks.configureTaskRegistryMaintenance,
   startTaskRegistryMaintenance: mocks.startTaskRegistryMaintenance,
   getInspectableActiveTaskRestartBlockers: mocks.getInspectableActiveTaskRestartBlockers,
@@ -346,4 +359,95 @@ describe("startGatewayEarlyRuntime", () => {
     expect(discoveryParams.gatewayDiscoveryServices).toEqual([service]);
     expect(discoveryParams.pluginRuntimeClaim).toBe(input.pluginRuntimeClaim);
   });
+});
+
+describe("early startup task maintenance", () => {
+  let maintenance: typeof import("../tasks/task-registry.maintenance.js");
+
+  beforeEach(async () => {
+    const tasks = await vi.importActual<typeof import("../tasks/runtime-internal.js")>(
+      "../tasks/runtime-internal.js",
+    );
+    maintenance = await vi.importActual<typeof import("../tasks/task-registry.maintenance.js")>(
+      "../tasks/task-registry.maintenance.js",
+    );
+    mocks.ensureTaskRuntimeStateReady.mockImplementation(tasks.ensureTaskRuntimeStateReady);
+    mocks.configureTaskRegistryMaintenance.mockImplementation(
+      maintenance.configureTaskRegistryMaintenance,
+    );
+    mocks.startTaskRegistryMaintenance.mockImplementation(maintenance.startTaskRegistryMaintenance);
+    mocks.setSkillsRemoteRegistry.mockReset();
+    mocks.registerSkillsChangeListener.mockReturnValue(() => {});
+  });
+
+  afterEach(async () => {
+    maintenance.stopTaskRegistryMaintenance();
+    vi.useRealTimers();
+    resetDetachedTaskLifecycleRuntimeForTests();
+    maintenance.resetTaskRegistryMaintenanceRuntimeForTests();
+    resetTaskRegistryForTests({ persist: false });
+    resetTaskFlowRegistryForTests({ persist: false });
+    await drainGlobalSingletonLifecycleState("close");
+  });
+  it.each([false, true])(
+    "preserves copied tasks only in an update canary (updateCanary: %s)",
+    async (updateCanary) => {
+      await withStateDirEnv("openclaw-canary-tasks-", async () => {
+        resetTaskRegistryForTests({ persist: false });
+        resetTaskFlowRegistryForTests({ persist: false });
+        vi.useFakeTimers();
+        const recoverTask = vi.fn(async () => ({ recovered: false }));
+        setDetachedTaskLifecycleRuntime({
+          ...getDetachedTaskLifecycleRuntime(),
+          tryRecoverTaskBeforeMarkLost: recoverTask,
+        });
+        const staleAt = Date.now() - 45 * 60_000;
+        const copiedTask = createTaskFixture("cli", {
+          task: "Synthetic copied task",
+          runId: "canary-copied-run",
+          lastEventAt: staleAt,
+          notifyPolicy: "silent",
+        });
+        const expiredTask = createTaskFixture("cli", {
+          task: "Synthetic expired task",
+          status: "succeeded",
+          lastEventAt: staleAt,
+          cleanupAfter: staleAt,
+          notifyPolicy: "silent",
+        });
+        const recentTask = createTaskFixture("cli", {
+          task: "Synthetic task within recovery grace",
+          runId: "canary-recent-run",
+          lastEventAt: Date.now() - 4 * 60_000,
+          notifyPolicy: "silent",
+        });
+        const earlyRuntime = await startGatewayEarlyRuntime(
+          earlyRuntimeInput({ minimalTestGateway: false, updateCanary }),
+        );
+        try {
+          // Exercise both the startup sweep and the recurring maintenance sweep.
+          for (const elapsedMs of [5_000, 60_000]) {
+            await vi.advanceTimersByTimeAsync(elapsedMs);
+            if (updateCanary) {
+              expect(getTaskById(copiedTask.taskId)).toEqual(copiedTask);
+              expect(getTaskById(expiredTask.taskId)).toEqual(expiredTask);
+              expect(getTaskById(recentTask.taskId)).toEqual(recentTask);
+              expect(recoverTask).not.toHaveBeenCalled();
+            } else {
+              expect(getTaskById(copiedTask.taskId)?.status).toBe("lost");
+              expect(getTaskById(expiredTask.taskId)).toBeUndefined();
+              expect(getTaskById(recentTask.taskId)?.status).toBe(
+                elapsedMs === 5_000 ? "running" : "lost",
+              );
+              expect(recoverTask).toHaveBeenCalledTimes(elapsedMs === 5_000 ? 1 : 2);
+            }
+          }
+        } finally {
+          maintenance.stopTaskRegistryMaintenance();
+          await earlyRuntime.skillsChangeUnsub();
+          vi.useRealTimers();
+        }
+      });
+    },
+  );
 });
